+++ code/source/OMIP.cpp
@@ -33,17 +33,16 @@ OMIP::OMIP(const MIP& otherMIP) : MIP(otherMIP) {
 
 OMIP& OMIP::updateBudgetConstr(double rhs) {
 	removeRow(getNumRows() - 1);
-	addBudgetConstr(rhs);
+	/// FIXED: Bug#8ccd9f4f2accb180c35b60b7a41880fe25fb38da  -- Negligible small value were passed as number different from 0.
+	addBudgetConstr( abs(rhs)<=EPSILON ? 0.0: rhs); 
 	return *this;
 }
 
 double OMIP::getSlackSum() {
 	std::vector<double> xStar = getSol();
-	/// FIXED: Bug#62f2110b0d2547498c59d7c19c3490ac15330119  -- Eliminated negligible value from the solution.
-	std::transform(xStar.begin() + getMIPNumVars(),xStar.end(), xStar.begin() + getMIPNumVars(), [](double d) -> double {return (abs(d)<EPSILON? 0.0 : d); });
-
-	double sum = std::accumulate(xStar.begin() + getMIPNumVars(), xStar.end(), 0.0); 
 	/// FIXED: Bug#5c77b0d838cf9df00715d2bae81ef822eb7ddbd5  -- Unexpected cast to int if init = 0.
+	double sum = std::accumulate(xStar.begin() + getMIPNumVars(), xStar.end(), 0.0); 
+	
 	if(sum < -EPSILON) 
 		throw MIPException(MIPEx::OutOfBound, "Negative value obtained! "+ std::to_string(sum));
 	return sum;