+++ Changelog.md
@@ -9,11 +9,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Reduced computational overhead  
 
 
+# [1.0.5] - 2025-04-01
+### Fixed
+- Discretized negligible values in `OMIP::updateBudgetConstr` function ([OMIP.cpp](code/source/OMIP.cpp)).
+- Discretized negligible values in the solution returned by `MIP::getObjFunction` function ([MIP.cpp](code/source/MIP.cpp)).
+- Discretized negligible values in the solution returned by `MIP::getSol` function ([MIP.cpp](code/source/MIP.cpp)).
+
+
 # [1.0.4] - 2025-03-31
 ### Fixed
-- Resolved unexpected behavior in `OMIP::getSlackSum` function ([OMIP.cpp](code/source/OMIP.cpp)).
-- Resolved error related to `MIP::MIP_INT_TOL` (MIP Integrality tolerance) in CPLEX optimization ([MIP.hpp](code/include/MIP.hpp))
-- Resolved error related to `MIP::MIP_SIMPLEX_FEAS_TOL` (MIP Simplex Feasibility tolerance) in CPLEX optimization ([MIP.hpp](code/include/MIP.hpp))
+- Fixed issue with `MIP::MIP_INT_TOL` (MIP Integrality Tolerance) in CPLEX optimization ([MIP.hpp](code/include/MIP.hpp)).
+- Fixed issue with `MIP::MIP_SIMPLEX_FEAS_TOL` (MIP Simplex Feasibility Tolerance) in CPLEX optimization ([MIP.hpp](code/include/MIP.hpp)).
 
 
 # [1.0.3] - 2025-03-30